--- a/kerchunk/hdf.py
+++ b/kerchunk/hdf.py
@@ -103,12 +103,6 @@ class SingleHdf5ToZarr:
             self._h5f = h5f
             fs, path = fsspec.core.url_to_fs(url, **(storage_options or {}))
             self.input_file = fs.open(path, "rb")
-        elif isinstance(h5f, h5py.Dataset):
-            group = h5f.file.create_group(f"{h5f.name}_")
-            group[f"{h5f.name}__"] = h5f
-            self._h5f = group
-            fs, path = fsspec.core.url_to_fs(url, **(storage_options or {}))
-            self.input_file = fs.open(path, "rb")
         else:
             raise ValueError("type of input `h5f` not recognised")
         self.spec = spec