@valeriupredoi 2024-02-21
Hi @martindurant  very many thanks for starting to look into this so quickly in a PR, mate :beer: 

Your code change addresses one issue (in terms of performance): the kerchunking/translation, and does so, in my test case, by reducing the runtime from about 100s to 75s - which is great, but not ideal :grin: The main problem here is `visititems()` that takes the bulk of time, and that's done on the entire HDF5/netCDF4 file. Here is a very, very rough POC of how this issue of `visititems` can be addressed: in `translate(self)`, when it comes to visiting items (creating the B-tree), we should do that just for the bit of HDF5 that we need to look at ie just the Dataset of interest - I could not think of a better way to move from Dataset to HDF5 object, but a nasty write-to-disk, but surely there are more elegant ways to convert a Dataset to a HDF5 in memory:

```python
        if self.var_pattern and self._h5f[self.var_pattern]:
            dset = self._h5f[self.var_pattern]
            self._h5f = h5py.File('data.hdf5', 'w')
            self._h5f.create_dataset(self.var_pattern, data=dset)
            self._h5f.visititems(self._translator)
 ```

this makes my visititems drop from 75s to 12s clearly with most of those 12s spent writing the silly file to disk :beer: