@valeriupredoi 2024-02-21
and indeed that ~12s is file IO, if one uses a cached `File` that time drops to about 1s just for `visititems` :+1: