@valeriupredoi 2024-02-21
Hi @martindurant I've struggled with h5py all day today, and the best I came up with, in order to restrict visiting items only to the bits of the file one needs (ie only the needed Datasets/variables) is to create an empty Group and store those datasets in `translator()`:

```python
        lggr.debug("Translation begins")
        self._transfer_attrs(self._h5f, self._zroot)
        if self.var and self._h5f[self.var]:
            if isinstance(self._h5f[self.var], h5py.Dataset):
                import time
                t1 = time.time()
                self._h5f.create_group(self.var + "_adhoc")
                self._h5f[self.var + "_adhoc"][self.var] = self._h5f[self.var]
                self._h5f = self._h5f[self.var + "_adhoc"]
                self._h5f.visititems(self._translator)
                t2 = time.time()
                print("New group creation and visititems took:", t2 - t1)
            else:
                self._h5f = dset  # already a Group
                self._h5f.visititems(self._translator)
        else:
            self._h5f.visititems(self._translator)
 ```
 
 thing is, this is ugly (because I don't really know HDF5 all too well, sorry), and not particularly fast, it is, in fact, slower than writing out a file, though we don't want to transfer any data at all. I hope to High Heavens and trust you have a much more efficient and elegant solution :smiley:  :beers: