@martindurant 2024-02-23
> storage_options['default_cache_type'] = "none"

Type "first" is usually the best option for HDF5.