@valeriupredoi 2024-02-23
whoa and a difference it makes, cheers muchly, Martin! Have a look at these numbers (for the same test above):
```
Normal Kerchunk with default cache "first"
------------------------------------------
Visititems took: 1.6179678440093994
Time to Translate and Dump Kerchunks to json file 2.377903461456299
Visititems took: 1.6005499362945557
Time to Translate and Dump Kerchunks to json file 2.2975118160247803
Visititems took: 1.6153967380523682
Time to Translate and Dump Kerchunks to json file 2.6384167671203613
Visititems took: 1.5885121822357178
Time to Translate and Dump Kerchunks to json file 2.504279136657715

Restricted to one Group/Dataset Kerchunk with default cache "first"
-------------------------------------------------------------------
Visititems took: 0.10222649574279785
Time to Translate and Dump Kerchunks to json file 0.7997150421142578
Visititems took: 0.10846853256225586
Time to Translate and Dump Kerchunks to json file 0.7331216335296631
Visititems took: 0.11911702156066895
Time to Translate and Dump Kerchunks to json file 0.82962965965271
Visititems took: 0.10615754127502441
Time to Translate and Dump Kerchunks to json file 0.8147380352020264
```