@martindurant 2024-02-23
I tried to adapt your version in the latest commit.