@martindurant 2024-02-27
So you're suggesting removing the Dataset possibility?