@valeriupredoi 2024-02-27
> So you're suggesting removing the Dataset possibility?

omdeed, I think it's too much of a headache to make that work at your end, and as far as I can see it works well at my end (user's end), so prob best to turn it off and only leave the `Group` input possibility? :beer: