@valeriupredoi 2024-02-28
attaching the file, so it's more readable
[test_file.json](https://github.com/fsspec/kerchunk/files/14435188/test_file.json)