@martindurant 2024-02-28
Is this different behaviour than without filtering the HDF?