@martindurant 2024-02-28
Also, since it's just JSON: can you edit out the offending filter and see if that's a fix?